Rename package to 'Kugelblitz'
