+++ Kugelblitz-Project-intellij-code-style.xml
@@ -1,4 +1,4 @@
-<code_scheme name="Entity-WS-Project">
+<code_scheme name="Kugelblitz-Project">
   <option name="GENERATE_FINAL_LOCALS" value="true" />
   <option name="GENERATE_FINAL_PARAMETERS" value="true" />
   <option name="USE_EXTERNAL_ANNOTATIONS" value="true" />
