+++ src/main/java/org/qubitpi/kugelblitz/SwaggerConfig.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.qubitpi.ws.entity;
+package org.qubitpi.kugelblitz;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
