+++ src/main/java/org/qubitpi/kugelblitz/arango/CollectionService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.qubitpi.ws.entity.arango;
+package org.qubitpi.kugelblitz.arango;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
