+++ src/test/java/org/qubitpi/kugelblitz/SwaggerIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.qubitpi.ws.entity;
+package org.qubitpi.kugelblitz;
 
 import static org.assertj.core.api.Assertions.assertThat;
